(culling) Extract stats panel to StatsPanel.jsx
Move CullingStats, CullingExcluded and EXCLUDED_LABELS verbatim out of
the shell into components/culling/StatsPanel.jsx. Pure move — the shell
imports CullingStats and renders it exactly as before.
